+++ config/routes.rb
@@ -95,6 +95,9 @@ Rails.application.routes.draw do
     end
 
     resources :auth, only: %i[create]
+    namespace :characters do
+      resources :import, only: %i[create]
+    end
     resources :characters, only: %i[index show destroy] do
       resources :notes, only: %i[index create update destroy], module: 'characters'
       resources :resources, only: %i[create update destroy], module: 'characters'