+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Cosmere PDF
+- importing data from Pathbuilder
 
 ### Modified
 - PF2 PDF